Move definition of `ShufflingOptions` to a separate file.
The same object will be used by different implemetnations of Reader, so needed to factor out.
