+++ petastorm/reader.py
@@ -21,6 +21,7 @@ import six
 from pyarrow import parquet as pq
 from six.moves.urllib.parse import urlparse
 
+import petastorm.shuffle_options
 from petastorm.cache import NullCache
 from petastorm.etl import dataset_metadata, rowgroup_indexing
 from petastorm.fs_utils import FilesystemResolver
@@ -34,26 +35,9 @@ from petastorm.workers_pool.ventilator import ConcurrentVentilator
 
 logger = logging.getLogger(__name__)
 
-
-class ShuffleOptions(object):
-    shuffle_row_groups = None
-    shuffle_row_drop_partitions = None
-
-    def __init__(self, shuffle_row_groups=True, shuffle_row_drop_partitions=1):
-        """
-        Constructor.
-
-        :param shuffle_row_groups: Whether to shuffle row groups (the order in which full row groups are read)
-        :param shuffle_row_drop_partitions: This is is a positive integer which determines how many partitions to
-            break up a row group into for increased shuffling in exchange for worse performance (extra reads).
-            For example if you specify 2 each row group read will drop half of the rows within every row group and
-            read the remaining rows in separate reads. It is recommended to keep this number below the regular row
-            group size in order to not waste reads which drop all rows.
-        """
-        if not isinstance(shuffle_row_drop_partitions, int) or not shuffle_row_drop_partitions >= 1:
-            raise ValueError('shuffle_row_drop_ratio must be positive integer')
-        self.shuffle_row_groups = shuffle_row_groups
-        self.shuffle_row_drop_partitions = shuffle_row_drop_partitions
+# For backward compatibility. Would clean this up with package reorganization in the future
+# https://github.com/uber/petastorm/issues/134
+ShuffleOptions = petastorm.shuffle_options.ShuffleOptions
 
 
 class Reader(object):
@@ -163,7 +147,7 @@ class Reader(object):
                 shuffle = True
             else:
                 logger.warning('shuffle option is deprecated. Please use shuffle_options instead')
-            shuffle_options = ShuffleOptions(shuffle)
+            shuffle_options = petastorm.shuffle_options.ShuffleOptions(shuffle)
         self._normalize_shuffle_options(shuffle_options, self.dataset)
         ventilator = self._create_ventilator(filtered_row_group_indexes, shuffle_options, num_epochs, worker_predicate)
